검색 TestCode 미비한 부분 수정
