+++ app/src/main/java/com/android/code/models/marvel/MarvelResponse.kt
@@ -17,72 +17,14 @@ data class SampleResponse(
 )
 
 data class MarvelResult(
-    @SerializedName("comics")
-    val comics: Comics,
     @SerializedName("description")
     val description: String,
-    @SerializedName("events")
-    val events: Events,
     @SerializedName("id")
     val id: Int,
-    @SerializedName("modified")
-    val modified: String,
     @SerializedName("name")
     val name: String,
-    @SerializedName("resourceURI")
-    val resourceURI: String,
-    @SerializedName("series")
-    val series: Series,
-    @SerializedName("stories")
-    val stories: Stories,
     @SerializedName("thumbnail")
-    val thumbnail: Thumbnail,
-    @SerializedName("urls")
-    val urls: List<Url>?
-)
-
-data class Comics(
-    @SerializedName("available")
-    val available: Int,
-    @SerializedName("collectionURI")
-    val collectionURI: String,
-    @SerializedName("items")
-    val items: List<ComicsMetaData>?,
-    @SerializedName("returned")
-    val returned: Int
-)
-
-data class Events(
-    @SerializedName("available")
-    val available: Int,
-    @SerializedName("collectionURI")
-    val collectionURI: String,
-    @SerializedName("items")
-    val items: List<EventsMetaData>?,
-    @SerializedName("returned")
-    val returned: Int
-)
-
-data class Series(
-    @SerializedName("available")
-    val available: Int,
-    @SerializedName("collectionURI")
-    val collectionURI: String,
-    @SerializedName("items")
-    val items: List<SeriesMetaData>,
-    @SerializedName("returned")
-    val returned: Int
-)
-
-data class Stories(
-    @SerializedName("available")
-    val available: Int,
-    @SerializedName("collectionURI")
-    val collectionURI: String,
-    @SerializedName("items")
-    val items: List<StoryMetaData>,
-    @SerializedName("returned")
-    val returned: Int
+    val thumbnail: Thumbnail?,
 )
 
 data class Thumbnail(
@@ -90,41 +32,4 @@ data class Thumbnail(
     val extension: String,
     @SerializedName("path")
     val path: String
-)
-
-data class Url(
-    @SerializedName("type")
-    val type: String,
-    @SerializedName("url")
-    val url: String
-)
-
-data class ComicsMetaData(
-    @SerializedName("name")
-    val name: String,
-    @SerializedName("resourceURI")
-    val resourceURI: String
-)
-
-data class EventsMetaData(
-    @SerializedName("name")
-    val name: String,
-    @SerializedName("resourceURI")
-    val resourceURI: String
-)
-
-data class SeriesMetaData(
-    @SerializedName("name")
-    val name: String,
-    @SerializedName("resourceURI")
-    val resourceURI: String
-)
-
-data class StoryMetaData(
-    @SerializedName("name")
-    val name: String,
-    @SerializedName("resourceURI")
-    val resourceURI: String,
-    @SerializedName("type")
-    val type: String
 )
